# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Library Management System API - Java Spring Boot Backend
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
  service?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface LoginResponse {
  token: string;
  refreshToken: string;
  email: string;
  name: string;
  role: string;
  userId: number;
}

export type RegisterRequestRole =
  (typeof RegisterRequestRole)[keyof typeof RegisterRequestRole];

export const RegisterRequestRole = {
  ROLE_ADMIN: "ROLE_ADMIN",
  ROLE_LIBRARIAN: "ROLE_LIBRARIAN",
  ROLE_FACULTY: "ROLE_FACULTY",
  ROLE_COLLEGE_STUDENT: "ROLE_COLLEGE_STUDENT",
  ROLE_SCHOOL_STUDENT: "ROLE_SCHOOL_STUDENT",
  ROLE_GENERAL_PATRON: "ROLE_GENERAL_PATRON",
  ROLE_STUDENT: "ROLE_STUDENT",
  ROLE_MEMBER: "ROLE_MEMBER",
} as const;

export interface RegisterRequest {
  email: string;
  password: string;
  name: string;
  /** @nullable */
  phone?: string | null;
  role: RegisterRequestRole;
  /** @nullable */
  studentId?: string | null;
  /** @nullable */
  collegeName?: string | null;
  /** @nullable */
  schoolGrade?: string | null;
  /** @nullable */
  parentEmail?: string | null;
  /** @nullable */
  membershipExpiry?: string | null;
}

export interface UserResponse {
  id: number;
  email: string;
  name: string;
  /** @nullable */
  phone?: string | null;
  role: string;
  /** @nullable */
  studentId?: string | null;
  /** @nullable */
  collegeName?: string | null;
  /** @nullable */
  schoolGrade?: string | null;
  enabled: boolean;
  createdAt?: string;
}

export type BookRequestMainCategory =
  (typeof BookRequestMainCategory)[keyof typeof BookRequestMainCategory];

export const BookRequestMainCategory = {
  COMPETITIVE_EXAM: "COMPETITIVE_EXAM",
  COLLEGE: "COLLEGE",
  SCHOOL: "SCHOOL",
  COMIC: "COMIC",
  HISTORY: "HISTORY",
  NON_FICTION: "NON_FICTION",
  FICTION: "FICTION",
  OTHER: "OTHER",
  TECHNOLOGY: "TECHNOLOGY",
  SCIENCE: "SCIENCE",
  ARTS: "ARTS",
  PHILOSOPHY: "PHILOSOPHY",
  LITERATURE: "LITERATURE",
  BIOGRAPHY: "BIOGRAPHY",
  GEOGRAPHY: "GEOGRAPHY",
  LAW: "LAW",
  BUSINESS: "BUSINESS",
  MEDICINE: "MEDICINE",
  TRAVEL: "TRAVEL",
  REFERENCE: "REFERENCE",
} as const;

export interface BookRequest {
  title: string;
  author: string;
  /** @nullable */
  isbn?: string | null;
  mainCategory: BookRequestMainCategory;
  subCategory: string;
  /** @nullable */
  publicationYear?: number | null;
  /** @nullable */
  publisher?: string | null;
  /** @nullable */
  edition?: string | null;
  /** @minimum 1 */
  totalCopies: number;
  /** @nullable */
  locationShelf?: string | null;
  /** @nullable */
  language?: string | null;
  referenceOnly?: boolean;
  /** @nullable */
  description?: string | null;
}

export interface BookResponse {
  id: number;
  title: string;
  author: string;
  /** @nullable */
  isbn?: string | null;
  mainCategory: string;
  mainCategoryDisplay: string;
  subCategory: string;
  /** @nullable */
  publicationYear?: number | null;
  /** @nullable */
  publisher?: string | null;
  /** @nullable */
  edition?: string | null;
  totalCopies: number;
  availableCopies: number;
  /** @nullable */
  locationShelf?: string | null;
  /** @nullable */
  language?: string | null;
  referenceOnly: boolean;
  /** @nullable */
  description?: string | null;
}

export interface BookPage {
  content: BookResponse[];
  totalElements: number;
  totalPages: number;
  number: number;
  size: number;
}

export interface CategoryNode {
  main: string;
  displayName: string;
  subcategories: string[];
}

export interface CategoryTree {
  categories: CategoryNode[];
}

export interface CheckoutRequest {
  bookId: number;
  userId: number;
}

export type LoanResponseStatus =
  (typeof LoanResponseStatus)[keyof typeof LoanResponseStatus];

export const LoanResponseStatus = {
  ACTIVE: "ACTIVE",
  RETURNED: "RETURNED",
  OVERDUE: "OVERDUE",
  RENEWED: "RENEWED",
} as const;

export interface LoanResponse {
  id: number;
  bookId: number;
  bookTitle: string;
  bookAuthor: string;
  userId: number;
  userName: string;
  checkoutDate: string;
  dueDate: string;
  /** @nullable */
  returnedDate?: string | null;
  fineAmount: number;
  finePaid: boolean;
  renewedCount: number;
  status: LoanResponseStatus;
}

export interface LoanPage {
  content: LoanResponse[];
  totalElements: number;
  totalPages: number;
  number: number;
  size: number;
}

export interface ReservationRequest {
  bookId: number;
}

export type ReservationResponseStatus =
  (typeof ReservationResponseStatus)[keyof typeof ReservationResponseStatus];

export const ReservationResponseStatus = {
  PENDING: "PENDING",
  READY: "READY",
  FULFILLED: "FULFILLED",
  EXPIRED: "EXPIRED",
  CANCELLED: "CANCELLED",
} as const;

export interface ReservationResponse {
  id: number;
  bookId: number;
  bookTitle: string;
  bookAuthor: string;
  userId: number;
  userName: string;
  status: ReservationResponseStatus;
  /** @nullable */
  expiryDate?: string | null;
  queuePosition: number;
  createdAt?: string;
}

export type DashboardStatsCategoryStatsItem = { [key: string]: unknown };

export type DashboardStatsRecentActivityItem = { [key: string]: unknown };

export interface DashboardStats {
  totalBooks: number;
  availableBooks: number;
  totalUsers: number;
  activeLoans: number;
  overdueLoans: number;
  pendingReservations: number;
  categoryStats: DashboardStatsCategoryStatsItem[];
  recentActivity: DashboardStatsRecentActivityItem[];
}

export interface RecommendationRequest {
  title: string;
  author: string;
  /** @nullable */
  isbn?: string | null;
  /** @nullable */
  reason?: string | null;
}

export interface RecommendationResponse {
  id: number;
  title: string;
  author: string;
  /** @nullable */
  isbn?: string | null;
  /** @nullable */
  reason?: string | null;
  status: string;
  userId: number;
  userName: string;
  createdAt?: string;
}

export type ChangeRoleRequestRole =
  (typeof ChangeRoleRequestRole)[keyof typeof ChangeRoleRequestRole];

export const ChangeRoleRequestRole = {
  ROLE_ADMIN: "ROLE_ADMIN",
  ROLE_LIBRARIAN: "ROLE_LIBRARIAN",
  ROLE_FACULTY: "ROLE_FACULTY",
  ROLE_COLLEGE_STUDENT: "ROLE_COLLEGE_STUDENT",
  ROLE_SCHOOL_STUDENT: "ROLE_SCHOOL_STUDENT",
  ROLE_GENERAL_PATRON: "ROLE_GENERAL_PATRON",
  ROLE_STUDENT: "ROLE_STUDENT",
  ROLE_MEMBER: "ROLE_MEMBER",
} as const;

export interface ChangeRoleRequest {
  role: ChangeRoleRequestRole;
}

export interface AuditLogResponse {
  id: number;
  /** @nullable */
  userId?: number | null;
  action: string;
  /** @nullable */
  entityType?: string | null;
  /** @nullable */
  entityId?: number | null;
  /** @nullable */
  description?: string | null;
  timestamp: string;
}

export interface ApiError {
  timestamp?: string;
  status: number;
  error: string;
  message: string;
  path?: string;
}

export type SearchBooksParams = {
  q?: string;
  category?: string;
  subcategory?: string;
  page?: number;
  size?: number;
};

export type GetAllLoansParams = {
  page?: number;
  size?: number;
};