# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Library Management System API - Java Spring Boot Backend
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user and get JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        refreshToken:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
        userId:
          type: integer
      required:
        - token
        - refreshToken
        - email
        - name
        - role
        - userId
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        name:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [ROLE_ADMIN, ROLE_LIBRARIAN, ROLE_FACULTY, ROLE_COLLEGE_STUDENT, ROLE_SCHOOL_STUDENT, ROLE_GENERAL_PATRON, ROLE_STUDENT, ROLE_MEMBER]
        studentId:
          type: ["string", "null"]
        collegeName:
          type: ["string", "null"]
        schoolGrade:
          type: ["string", "null"]
        parentEmail:
          type: ["string", "null"]
        membershipExpiry:
          type: ["string", "null"]
          format: date
      required:
        - email
        - password
        - name
        - role
    UserResponse:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
        studentId:
          type: ["string", "null"]
        collegeName:
          type: ["string", "null"]
        schoolGrade:
          type: ["string", "null"]
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - role
        - enabled
...
    ApiError:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
      required:
        - status
        - error
        - message