# lib/api-zod/src/generated/types/loginResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Library Management System API - Java Spring Boot Backend
 * OpenAPI spec version: 1.0.0
 */

export interface LoginResponse {
  token: string;
  refreshToken: string;
  email: string;
  name: string;
  role: string;
  userId: number;
}