# lib/api-zod/src/generated/types/registerRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Library Management System API - Java Spring Boot Backend
 * OpenAPI spec version: 1.0.0
 */
import type { RegisterRequestRole } from "./registerRequestRole";

export interface RegisterRequest {
  email: string;
  password: string;
  name: string;
  /** @nullable */
  phone?: string | null;
  role: RegisterRequestRole;
  /** @nullable */
  studentId?: string | null;
  /** @nullable */
  collegeName?: string | null;
  /** @nullable */
  schoolGrade?: string | null;
  /** @nullable */
  parentEmail?: string | null;
  /** @nullable */
  membershipExpiry?: Date | null;
}