# lib/api-zod/src/generated/types/userResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Library Management System API - Java Spring Boot Backend
 * OpenAPI spec version: 1.0.0
 */

export interface UserResponse {
  id: number;
  email: string;
  name: string;
  /** @nullable */
  phone?: string | null;
  role: string;
  /** @nullable */
  studentId?: string | null;
  /** @nullable */
  collegeName?: string | null;
  /** @nullable */
  schoolGrade?: string | null;
  enabled: boolean;
  createdAt?: Date;
}